requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.27.0.post1  # https://github.com/encode/uvicorn



# Django
# ------------------------------------------------------------------------------
django==4.2.10  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-redis==5.4.0  # https://github.com/jazzband/django-redis

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
djangorestframework-simplejwt[crypto]==5.3.0  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html
dj-rest-auth==5.0.2  # https://dj-rest-auth.readthedocs.io/en/latest/index.html

# DRF-spectacular for api documentation
drf-spectacular==0.27.1  # https://github.com/tfranzel/drf-spectacular



# Django Admin
# ------------------------------------------------------------------------------
django-admin-honeypot1==1.2.0  # https://pypi.org/project/django-admin-honeypot1/
django-jet-reboot==1.3.7  # https://pypi.org/project/django-jet-reboot/
feedparser==6.0.11  # https://feedparser.readthedocs.io/en/latest/
google-api-python-client==1.4.1  # https://pypi.org/project/django-admin-honeypot1/



# Custom Utilities
# ------------------------------------------------------------------------------
# log informations
loguru==0.7.2
# minify html
django-htmlmin==0.11.0
